requirements.txt
allure-pytest==2.15.0
allure-python-commons==2.15.0
ansi2html==1.9.2
attrs==25.4.0
certifi==2025.10.5
charset-normalizer==3.4.4
colorama==0.4.6
docutils==0.22.3
et_xmlfile==2.0.0
execnet==2.1.1
Faker==37.12.0
Genshi==0.7.9
greenlet==3.2.4
htmlmin2==0.1.13
idna==3.11
iniconfig==2.3.0
Jinja2==3.1.6
MarkupSafe==3.0.3
numpy==2.3.4
openpyxl==3.1.5
packaging==25.0
pandas==2.3.3
playwright==1.55.0
pluggy==1.6.0
pyee==13.0.0
Pygments==2.19.2
pytest==8.4.2
pytest-base-url==2.1.0
pytest-html==4.1.1
pytest-metadata==3.1.1
pytest-playwright==0.7.1
pytest-reporter==0.5.3
pytest-reporter-html1==0.9.5
pytest-xdist==3.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-json-logger==4.0.0
python-slugify==8.0.4
pytz==2025.2
requests==2.32.5
rst2html5==2.0.1
six==1.17.0
SQLAlchemy==2.0.44
sqlparse==0.5.3
text-unidecode==1.3
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
xlrd==2.0.2
zipp==3.23.0
